# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Timefly API
  description: MongoDB ExpressJS NodeJS REST API for a Project Management App.<br />Github repository available [here](https://github.com/aneguet/MEVN_ProjectManagementApp).<br />Some endpoints require `auth-token` in order to be executed.
  version: 1.0.0
servers:
  - url: https://timefly-mevn.herokuapp.com/api/
    description: Heroku remote server
  - url: http://localhost:4000/api/
    description: Localhost
components:
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: auth-token
  schemas:
...
    Task:
      type: object
      required:
        - 'project_id'
        - 'name'
        - 'description'
        - 'hours_allocated'
      properties:
        project_id:
          type: string
        name:
          type: string
        description:
          type: string
        assigned_to:
          type: string
          description: 'Must be an id composed by numbers and letters. The task is assigned to the user creating the task by default.'
        hours_allocated:
          type: number
          default: 0
        hours_used:
          type: number
          default: 0
        task_status:
          type: string
          default: 'Todo'
          enum: [Todo, Doing, Done]
    Technology:
      type: object

      required:
        - 'tech_name'
      properties:
        tech_name:
          type: string
          example: 'PHP'
paths:
...
  /tasks:
    get:
      summary: Get all tasks (Admin functionality)
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Tasks successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        403:
          description: This request requires an Admin role
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tasks/byProjectAndUser:
    get:
      summary: Get all tasks by project ID and logged in user ID
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: project ID
          required: true
          type: string
      responses:
        200:
          description: Tasks successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tasks/byProject:
    get:
      summary: Get all tasks by project ID
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: project ID
          required: true
          type: string
      responses:
        200:
          description: Tasks successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tasks/task:
    get:
      summary: Get task by task ID
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: project ID
          required: true
          type: string
      responses:
        200:
          description: Task successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    post:
      summary: Create new task
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        200:
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Update task by ID
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: Task ID
          required: true
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        200:
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete task by ID
      tags:
        - Task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: Task ID
          required: true
          type: string
      responses:
        200:
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /technologies:
    get:
      summary: Get all technologies
      tags:
        - Technology
      responses:
        200:
          description: Technologies successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      summary: Create new technology
      tags:
        - Technology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Technology'
      responses:
        200:
          description: Technology successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /technologies/technology:
    get:
      summary: Get technology by ID
      tags:
        - Technology
      parameters:
        - name: id
          in: header
          description: Technology ID
          required: true
          type: string
      responses:
        200:
          description: Technology successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete technology by ID
      tags:
        - Technology
      parameters:
        - name: id
          in: header
          description: Technology ID
          required: true
          type: string
      responses:
        200:
          description: Technology successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string